refactor TypedActionButton to new file & components
